Se modifica archivo original para resolución de problema Broken Calculator
